email_agent/requirements.txt
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
google-api-python-client==2.110.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
openai==1.3.7
langchain==0.0.350
langchain-core==0.1.3
langchain-text-splitters==0.0.1
faiss-cpu==1.7.4
sqlmodel==0.0.14
aiosqlite==0.19.0
apscheduler==3.10.4
python-dotenv==1.0.0
pydantic==2.5.0
pandas==2.1.4
numpy==1.24.4
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0